Refactor: 상위 스튜디오 소개 컴포넌트에 대해  StudioDescription으로 컴포넌트 분리
