+++ pages/components/StudioDanceClass.tsx
